aitu-masters/advanced-computer-networks-security/practice_4/requirements.txt
openapi-generator-cli
build
fastapi[standard]
uvicorn[standard]
pyjwt[crypto]